sync: smoother teams bulk uploading (fixes #16754) (#16772)
